pyproject.toml
[tool.poetry]
name = "icon-cli"
version = "0.1.0"
description = ""
authors = ["Brian Li <brian@brianli.com>"]

[tool.poetry.scripts]
icon-cli = "icon_cli.main:app"

[tool.poetry.dependencies]
python = "^3.8"
iconsdk = {extras = ["all"], version = "^1.3.4"}
rich = {extras = ["all"], version = "^10.4.0"}
PyYAML = "^5.4.1"
typer = "^0.3.2"
python-dotenv = "^0.18.0"
petl = "^1.7.4"
pandas = "^1.3.1"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {extras = ["all"], version = "^21.6b0"}
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
